+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.9.5] - 2021-03-12
+## [0.9.5] - 2021-03-11
 ### Changed
+- Updated Python project examples
+- Added integration request parameters for API Gateway
+- Added request validator for API Gateway
 - Added the ability to add `logs_expiration` parameter in lambda_config.json.
 
 ## [0.9.4] - 2020-10-16